requirements.txt
fastapi
uvicorn
pandas
python-multipart
groq
python-dotenv
email-validator
openpyxl
